requirements.txt
opensea==0.1.5
web3==5.25.0
pandas==1.3.4
etherscan-python==2.1.0
plotly_express==0.4.1
streamlit==1.2.0
python-igraph==0.9.8

